Lazy load platform sub routes.
