+++ src/smart-components/dialog-routes/index.js
@@ -1,9 +1,11 @@
 import React, { Suspense, Fragment } from 'react';
 import PortfolioRoutes from './portfolio-routes';
+import PlatformRoutes from './platform-routes';
 
 const DialogRoutes = () => (
   <Suspense fallback={Fragment}>
     <PortfolioRoutes />
+    <PlatformRoutes />
   </Suspense>
 );
